% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{$,AkoyaExperiment-method}
\alias{$,AkoyaExperiment-method}
\title{Access metadata columns with $}
\usage{
\S4method{$}{AkoyaExperiment}(x, name)
}
\arguments{
\item{x}{An \code{\link{AkoyaExperiment}} object.}

\item{name}{Column name.}
}
\value{
The requested metadata column.
}
\description{
Access metadata columns with $
}
