% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dim,SpatialCellData-method}
\alias{dim,SpatialCellData-method}
\title{Get Dimensions of a SpatialCellData Object}
\usage{
\S4method{dim}{SpatialCellData}(x)
}
\arguments{
\item{x}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\value{
Integer vector: number of cells, number of markers.
}
\description{
Get Dimensions of a SpatialCellData Object
}
