% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Markers}
\alias{Markers}
\alias{Markers,SpatialCellData-method}
\title{Get Marker Names}
\usage{
Markers(object)

\S4method{Markers}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\value{
Character vector.
}
\description{
Get Marker Names
}
