% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{Reductions}
\alias{Reductions}
\alias{Reductions,SpatialCellData-method}
\title{List Available Reductions}
\usage{
Reductions(object)

\S4method{Reductions}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link{SpatialCellData-class}} object.}
}
\value{
Character vector of reduction names.
}
\description{
List Available Reductions
}
