% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,SpatialCellData-method}
\alias{show,SpatialCellData-method}
\title{Show a SpatialCellData Object}
\usage{
\S4method{show}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\description{
Show a SpatialCellData Object
}
