% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[,SpatialCellData-method}
\alias{[[,SpatialCellData-method}
\title{Access Metadata Columns with [[}
\usage{
\S4method{[[}{SpatialCellData}(x, i)
}
\arguments{
\item{x}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}

\item{i}{Column name (character).}
}
\value{
The requested metadata column.
}
\description{
Access Metadata Columns with [[
}
