% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Coords}
\alias{Coords}
\alias{Coords,SpatialCellData-method}
\title{Get Spatial Coordinates}
\usage{
Coords(object)

\S4method{Coords}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\value{
Data frame with columns \code{x} and \code{y}.
}
\description{
Get Spatial Coordinates
}
