% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Coords}
\alias{Coords}
\alias{Coords,AkoyaExperiment-method}
\title{Get Spatial Coordinates}
<<<<<<< Updated upstream
\description{
Returns the spatial coordinates from an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.
}
=======
>>>>>>> Stashed changes
\usage{
Coords(object)

\S4method{Coords}{AkoyaExperiment}(object)
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\value{
Data frame with columns \code{x} and \code{y}.
}
\description{
Get Spatial Coordinates
}
