% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dim,AkoyaExperiment-method}
\alias{dim,AkoyaExperiment-method}
<<<<<<< Updated upstream
\title{Get Dimensions of an AkoyaExperiment}
\description{
Returns the dimensions (number of cells and number of markers) of an
\code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.
}
=======
\title{Get dimensions of an AkoyaExperiment}
>>>>>>> Stashed changes
\usage{
\S4method{dim}{AkoyaExperiment}(x)
}
\arguments{
<<<<<<< Updated upstream
  \item{x}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{x}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\value{
Integer vector: number of cells, number of markers.
}
\description{
Get dimensions of an AkoyaExperiment
}
