% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{Embeddings}
\alias{Embeddings}
\alias{Embeddings,SpatialCellData-method}
\title{Get a Dimensionality-Reduction Embedding}
\usage{
Embeddings(object, reduction = "pca")

\S4method{Embeddings}{SpatialCellData}(object, reduction = "pca")
}
\arguments{
\item{object}{A \code{\link{SpatialCellData-class}} object.}

\item{reduction}{Character. Name of the reduction (e.g. \code{"pca"},
\code{"umap"}).}
}
\value{
A numeric matrix (cells x dimensions).
}
\description{
Get a Dimensionality-Reduction Embedding
}
\examples{
data(phenoscapR_example)
Embeddings(RunPCA(phenoscapR_example, n_pcs = 5), "pca")[1:3, ]
}
