% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetData}
\alias{GetData}
\alias{GetData,AkoyaExperiment-method}
\title{Get Expression Data}
<<<<<<< Updated upstream
\description{
Retrieve raw counts or normalised data from an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}}.
}
=======
>>>>>>> Stashed changes
\usage{
GetData(object, slot = "data")

\S4method{GetData}{AkoyaExperiment}(object, slot = "data")
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
  \item{slot}{Character. \code{"counts"} for raw data or \code{"data"}
    (default) for normalised data.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}

\item{slot}{Character. \code{"counts"} for raw data or \code{"data"}
(default) for normalised data.}
>>>>>>> Stashed changes
}
\value{
A numeric matrix (cells x markers).
}
\description{
Retrieve raw counts or normalised data from an AkoyaExperiment.
}
