% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetData}
\alias{GetData}
\alias{GetData,SpatialCellData-method}
\title{Get Expression Data}
\usage{
GetData(object, slot = "data")

\S4method{GetData}{SpatialCellData}(object, slot = "data")
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}

\item{slot}{Character. \code{"counts"} for raw data or \code{"data"}
(default) for normalised data.}
}
\value{
A numeric matrix (cells x markers).
}
\description{
Retrieve raw counts or normalised data from a SpatialCellData object.
}
