% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Idents}
\alias{Idents}
\alias{Idents,AkoyaExperiment-method}
\title{Get Active Cell Identities}
\usage{
Idents(object)

\S4method{Idents}{AkoyaExperiment}(object)
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\value{
A character vector.
}
\description{
Returns the phenotype labels if set, otherwise sample identities.
}
