% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Idents}
\alias{Idents}
\alias{Idents,SpatialCellData-method}
\title{Get Active Cell Identities}
\usage{
Idents(object)

\S4method{Idents}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\value{
A character vector.
}
\description{
Returns the phenotype labels if set, otherwise sample identities.
}
