% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{InteractionPlot}
\alias{InteractionPlot}
\title{Plot Spatial Interaction Heatmap}
\usage{
InteractionPlot(interactions, title = NULL)
}
\arguments{
\item{interactions}{Data frame as returned by
\code{\link{InteractionMatrix}}.}

\item{title}{Character or \code{NULL}.}
}
\value{
A \code{ggplot} object.
}
\description{
Heatmap of pairwise spatial interaction scores between phenotypes.
}
\examples{
interactions <- data.frame(
  from = rep(c("CD3+", "CD8+"), each = 2),
  to = rep(c("CD3+", "CD8+"), 2),
  observed = c(50, 30, 25, 40),
  expected = rep(35, 4),
  interaction_score = log2(c(50, 30, 25, 40) / 35)
)
InteractionPlot(interactions)

}
