% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Markers}
\alias{Markers}
\alias{Markers,AkoyaExperiment-method}
\title{Get Marker Names}
<<<<<<< Updated upstream
\description{
Returns the marker names from an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.
}
=======
>>>>>>> Stashed changes
\usage{
Markers(object)

\S4method{Markers}{AkoyaExperiment}(object)
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\value{
Character vector.
}
\description{
Get Marker Names
}
