% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Meta}
\alias{Meta}
\alias{Meta,AkoyaExperiment-method}
\title{Get or Set Cell Metadata}
<<<<<<< Updated upstream
\description{
Returns the cell metadata data frame from an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.
}
=======
>>>>>>> Stashed changes
\usage{
Meta(object)

\S4method{Meta}{AkoyaExperiment}(object)
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\value{
Data frame of cell metadata.
}
\description{
Get or Set Cell Metadata
}
