% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Meta}
\alias{Meta}
\alias{Meta,SpatialCellData-method}
\title{Get or Set Cell Metadata}
\usage{
Meta(object)

\S4method{Meta}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\value{
Data frame of cell metadata.
}
\description{
Get or Set Cell Metadata
}
\examples{
data(phenoscapR_example)
head(Meta(phenoscapR_example))
}
