% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NCells}
\alias{NCells}
\alias{NCells,SpatialCellData-method}
\title{Get the Number of Cells}
\usage{
NCells(object)

\S4method{NCells}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\value{
Integer.
}
\description{
Get the Number of Cells
}
\examples{
data(phenoscapR_example)
NCells(phenoscapR_example)
}
