% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NCells}
\alias{NCells}
\alias{NCells,AkoyaExperiment-method}
\title{Get the Number of Cells}
<<<<<<< Updated upstream
\description{
Returns the number of cells in an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.
}
=======
>>>>>>> Stashed changes
\usage{
NCells(object)

\S4method{NCells}{AkoyaExperiment}(object)
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\value{
Integer.
}
\description{
Get the Number of Cells
}
