% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NMarkers}
\alias{NMarkers}
\alias{NMarkers,SpatialCellData-method}
\title{Get the Number of Markers}
\usage{
NMarkers(object)

\S4method{NMarkers}{SpatialCellData}(object)
}
\arguments{
\item{object}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}
}
\value{
Integer.
}
\description{
Get the Number of Markers
}
\examples{
data(phenoscapR_example)
NMarkers(phenoscapR_example)
}
