% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NMarkers}
\alias{NMarkers}
\alias{NMarkers,AkoyaExperiment-method}
\title{Get the Number of Markers}
<<<<<<< Updated upstream
\description{
Returns the number of markers in an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.
}
=======
>>>>>>> Stashed changes
\usage{
NMarkers(object)

\S4method{NMarkers}{AkoyaExperiment}(object)
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\value{
Integer.
}
\description{
Get the Number of Markers
}
