% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoscapR-package.R
\docType{package}
\name{phenoscapR-package}
\alias{phenoscapR}
\alias{phenoscapR-package}
\title{phenoscapR: Read, Analyse, and Visualise Akoya Spatial Biology Data}
\description{
Tools for reading, processing, analysing, and visualising multiplexed spatial biology data produced by Akoya Biosciences platforms (PhenoCycler, CODEX, PhenoImager). Provides functions to import cell segmentation outputs, perform quality control, normalise marker intensities, compute spatial statistics such as nearest neighbour distances and cell density, and generate publication-ready visualisations of cell phenotype maps and spatial interaction networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-heller/phenoscapR}
  \item Report bugs at \url{https://github.com/r-heller/phenoscapR/issues}
}

}
\author{
\strong{Maintainer}: phenoscapR contributors \email{phenoscapr@example.com}

}
\keyword{internal}
