% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,AkoyaExperiment-method}
\alias{show,AkoyaExperiment-method}
\title{Show an AkoyaExperiment}
<<<<<<< Updated upstream
\description{
Display a summary of an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object, including
the number of cells, samples, markers, normalisation status, phenotypes,
and spatial results.
}
=======
>>>>>>> Stashed changes
\usage{
\S4method{show}{AkoyaExperiment}(object)
}
\arguments{
<<<<<<< Updated upstream
  \item{object}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
=======
\item{object}{An \code{\link{AkoyaExperiment}} object.}
>>>>>>> Stashed changes
}
\description{
Show an AkoyaExperiment
}
