% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{SpatialClusters}
\alias{SpatialClusters}
\title{Spatial Clusters (AkoyaExperiment)}
\usage{
SpatialClusters(object, k, method = c("kmeans", "hierarchical"))
}
\arguments{
\item{object}{An \code{\link{AkoyaExperiment-class}} object.}

\item{k}{Integer. Number of clusters.}

\item{method}{Character. \code{"kmeans"} (default) or \code{"hierarchical"}.}
}
\value{
An \code{\link{AkoyaExperiment-class}} with a \code{cluster}
column in \code{meta_data}.
}
\description{
Clusters cells based on spatial coordinates.
}
\examples{
counts <- matrix(rnorm(100), nrow = 50,
                 dimnames = list(NULL, c("CD3", "CD8")))
coords <- data.frame(
  x = c(rnorm(25, 0, 5), rnorm(25, 50, 5)),
  y = c(rnorm(25, 0, 5), rnorm(25, 50, 5)))
obj <- CreateAkoyaObject(counts, coords)
obj <- SpatialClusters(obj, k = 2)
table(Meta(obj)$cluster)

}
