% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[,AkoyaExperiment-method}
\alias{[,AkoyaExperiment-method}
\title{Subset an AkoyaExperiment}
<<<<<<< Updated upstream
\description{
Subset an \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object by cell indices (rows)
and/or marker indices (columns).
}
=======
>>>>>>> Stashed changes
\usage{
\S4method{[}{AkoyaExperiment}(x, i, j, drop = FALSE)
}
\arguments{
<<<<<<< Updated upstream
  \item{x}{An \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}} object.}
  \item{i}{Cell indices (integer or logical).}
  \item{j}{Marker indices (integer, logical, or character).}
  \item{drop}{Ignored.}
}
\value{
A subsetted \code{\link[=AkoyaExperiment-class]{AkoyaExperiment}}.
=======
\item{x}{An \code{\link{AkoyaExperiment}} object.}

\item{i}{Cell indices (integer or logical).}

\item{j}{Marker indices (integer, logical, or character).}

\item{drop}{Ignored.}
}
\value{
A subsetted \code{AkoyaExperiment}.
>>>>>>> Stashed changes
}
\description{
Subset an AkoyaExperiment
}
