% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[,SpatialCellData-method}
\alias{[,SpatialCellData-method}
\title{Subset a SpatialCellData Object}
\usage{
\S4method{[}{SpatialCellData}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A \code{\link[=SpatialCellData-class]{SpatialCellData}} object.}

\item{i}{Cell indices (integer or logical).}

\item{j}{Marker indices (integer, logical, or character).}

\item{drop}{Ignored.}
}
\value{
A subsetted \code{SpatialCellData} object.
}
\description{
Subset a SpatialCellData Object
}
